requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-google-genai
langchain-community
tavily-python
python-dotenv
beautifulsoup4
requests